le08_11/requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.0
asynctest==0.13.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
frozenlist==1.2.0
idna==3.3
importlib-metadata==4.8.1
Markdown==3.3.4
multidict==5.2.0
mysqlclient==2.0.3
plotly==5.3.1
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
telepot==12.7
tenacity==8.0.1
twilio==7.3.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.2
zipp==3.6.0
